接近・到着閾値の計算に次の駅への距離を考慮
